Reliable Detection of the Tag16h5 family (#6)
* Consolidate shell scripts
Consolidated the shell scripts into one file with the caveat that the user cannot run the script with root permissions.
Updated the install directions in README.md to relflect this change.
Updated the .gitignore to reflect a coming change involving the way callibration images are stored.
Removed test.py as it serves no real function.

Signed-off-by: Alex Pereira <alex.pereira.6464@gmail.com>

* Logger and flips
Added Logger.py which will serve as a utility class to easily access the Python logger.
Updated all the other classes to pass debug and info data to the logger.
Updated main.py and calibration.py to flip the image ONLY before displaying the data to the user. This makes the display from cv.imshow() look like what a human would expect while still passing the correct data to the modules for processing.

Signed-off-by: Alex Pereira <alex.pereira.6464@gmail.com>

* Formatting
Decided that the libraries we installed with pip and ninja will be called packages and the Python code written here will be called modules.

Signed-off-by: Alex Pereira <alex.pereira.6464@gmail.com>

* Updated shell script
Renamed InstallPrerequisites.sh to Install.sh as this script will dowlad the code and install all dependencies correctly.

Signed-off-by: Alex Pereira <alex.pereira.6464@gmail.com>

* bestResult entry
Added the key for the bestResult entry.

Signed-off-by: Alex Pereira <alex.pereira.6464@gmail.com>

* Updated calibration method
Updated the calibration method to take a parameter which is the number of images we want taken.
Added a recursive call to calibrateCamera() that restarts the callibration if more than 2/3's of the images contined bad data.

Signed-off-by: Alex Pereira <alex.pereira.6464@gmail.com>

* Update README.md
Updated the README to include the most recent changes.
Reordered methods in pose_estimation.py by level of importance.

Signed-off-by: Alex Pereira <alex.pereira.6464@gmail.com>

* Jetson time
The Jetson is now capable of tracking its on time using the field's timer, therefore a few methods and entries in communications.py are not needed.

Signed-off-by: Alex Pereira <alex.pereira.6464@gmail.com>

* Init loggers
Added init loggers to all functional classes (excludes the utiliy classes).

Signed-off-by: Alex Pereira <alex.pereira.6464@gmail.com>

* Update to Install.sh
Updated Install.sh to reflect changes to the installation procedure.

Signed-off-by: Alex Pereira <alex.pereira.6464@gmail.com>

* Comments
Removed unecessary comments.

Signed-off-by: Alex Pereira <alex.pereira.6464@gmail.com>

* Reccalibration
Added logic to restart the calibration if very few chessboard were actually detected. This makes the caliration as acurate as possible.
Remade the method for creating calibration images from a video capture as this was adding color to the saved images, throwing off the calibration.

Signed-off-by: Alex Pereira <alex.pereira.6464@gmail.com>

* Filtering Noise and Removing False Positives (#1)
Added logic to take the hamming, decision margin, and error into account when finding valid tags. If a tag does not meet the strict filter criteria, it is thrown out as noise.
Removed atDetectTags() and all use of the offical AprilTags library as Duckie Town AprilTags provides far more info that can be used for elimating false poistives.
Removed draw() as it did the same thing as the previously existing draw methods.

Note: In order to elimate nosie, we lost the ability to see far away tags. The current reliable maximum is 7 feet, but that was for a 720p stream.
This can be adressed by fine tuning the filter's exclusion parameters.

Signed-off-by: Alex Pereira <alex.pereira.6464@gmail.com>

* Update calibration.py

Changed comments to match English standards.

Signed-off-by: Alex Pereira <alex.pereira.6464@gmail.com>

* Updated to Pupil Apriltags
Updated from dt-apriltags to pupil-apriltags to allow cross compatibilty for Windows. This also addresses the issue of dt-apriltags being three years out of date.
Replaced a period with a comma in calibration.py to allow the logger to correctly store information.

Signed-off-by: Alex Pereira <alex.pereira.6464@gmail.com>

* Similar indenting structure
Updated all files to have a similar indenting structure.

Signed-off-by: Alex Pereira <alex.pereira.6464@gmail.com>

* Massive Increase to Detection Distance
By fine tuning parameters in apriltags.py, the reliable detection distance has increased massively from 7 feet to over 30 feet.
Another filter has also been added to apriltags.py to remove any tag that is not present on the field (tags 1 - 8)

Signed-off-by: Alex Pereira <alex.pereira.6464@gmail.com>

* Making an issue address official (#5)
This issue was technically adressed in bf1dc2463cd83482cb1fe4e97216914cd4c858b1, but the issue number wasn't added to the commit header.

Signed-off-by: Alex Pereira <alex.pereira.6464@gmail.com>

* Changed all files to have similar indent structure

Signed-off-by: Alex Pereira <alex.pereira.6464@gmail.com>